requirements.txt
python-telegram-bot==13.7
pillow==9.2.0
boto3==1.28.38
requests==2.31.0
urllib3<2.0.0
python-dotenv==1.0.0
openai==1.79.0